% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_ping.R
\name{binance_ping}
\alias{binance_ping}
\title{Ping to Binance REST API}
\usage{
binance_ping(api)
}
\arguments{
\item{api}{Character. Reference API. If it is \code{missing}, the default, will be used \code{"spot"}. Available options are:
\itemize{
\item \code{"spot"}: for endpoint \href{https://binance-docs.github.io/apidocs/spot/en/#test-connectivity}{api/v3/ping}. The ip weight is 1.
\item \code{"fapi"}: for endpoint \href{https://binance-docs.github.io/apidocs/futures/en/#test-connectivity}{fapi/v1/ping}. The ip weight is 1.
\item \code{"dapi"}: for endpoint \href{https://binance-docs.github.io/apidocs/delivery/en/#test-connectivity}{dapi/v1/ping}. The ip weight is 1.
\item \code{"eapi"}: for endpoint \href{https://binance-docs.github.io/apidocs/voptions/en/#test-connectivity}{eapi/v1/ping}. The ip weight is 1.
}}
}
\value{
A logical value. It is \code{TRUE} if the connection was successful, otherwise it is \code{FALSE}.
}
\description{
Check the connection to the Binance API.
}
\examples{
# Test connection to spot api
binance_ping("spot")

# Test connection to futures usd-m api
binance_ping("fapi")

# Test connection to futures coin-m api
binance_ping("dapi")

# Test connection to options api
binance_ping("eapi")

}
\keyword{market}
\keyword{ping}
\keyword{rest}
