% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_trades.R
\name{binance_trades}
\alias{binance_trades}
\title{Binance Aggregated Historical Trades}
\usage{
binance_trades(pair, 
               api, 
               from,
               to,
               quiet = FALSE)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{api}{Character. Reference API. If it is \code{missing}, the default, will be used \code{"spot"}. Available options are:
\itemize{
\item \code{"spot"}: for endpoint \href{https://binance-docs.github.io/apidocs/spot/en/#old-trade-lookup}{api/v3/aggTrades}.
\item \code{"fapi"}: for endpoint \href{https://binance-docs.github.io/apidocs/futures/en/#old-trade-lookup}{fapi/v1/aggTrades}.
\item \code{"dapi"}: for endpoint \href{https://binance-docs.github.io/apidocs/delivery/en/#old-trade-lookup}{dapi/v1/aggTrades}.
\item \code{"eapi"}: for endpoint \href{https://binance-docs.github.io/apidocs/voptions/en/#old-trade-lookup}{eapi/v1/aggTrades}.
}}

\item{from}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. Start time for historical data.
If it is \code{missing}, the default, will be used as start date \code{Sys.time()-lubridate::minutes(10)}.}

\item{to}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. End time for historical data.
If it is \code{missing}, the default, will be used as end date \code{\link[=Sys.time]{Sys.time()}}.}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
A \code{\link[tibble]{tibble}} with 9 columns:
\itemize{
\item \code{date}: \code{\link[=POSIXt-class]{POSIXt}}, trade execution date;
\item \code{market}: Character, API.
\item \code{pair}: Character, trading pair.
\item \code{price}: Numeric, trade price.
\item \code{quantity}: Numeric, trade quantity.
\item \code{side}: Character, trade side. Can be \code{"BUY"} or \code{"SELL"}.
\item \code{agg_id}: Integer, aggregated trade id.
\item \code{first_id}: Integer, first trade id for aggregation.
\item \code{last_id}: Integer, last trade id for aggregation.
}
}
\description{
Get aggregated historical trades data for a trading pair.
}
\examples{

# Get trades in last 10 minutes for BTCUSDT
binance_trades(pair = "BTCUSDT", api = "spot", from = NULL, to = NULL)

# Get trades in last 10 minutes for LAZIOUSDT
binance_trades(pair = "LAZIOUSDT", api = "spot", from = "2023-01-01", to = "2023-01-02")

# Get trades in last 10 minutes for BTCUSDT
binance_trades(pair = "BTCUSDT", api = "fapi", from = NULL, to = NULL)

# Get trades in last 10 minutes for BTCUSD_PERP
binance_trades(pair = "BTCUSD_PERP", api = "dapi", from = NULL, to = NULL)

}
