% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binanceWebSocket.R
\name{binanceWebSocket}
\alias{binanceWebSocket}
\title{R6 Class for Binance Web Socket}
\description{
Binance web socket description
}
\details{
Binance web socket details
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{stream}}{Get all streams}

\item{\code{info}}{Get all streams info}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{binanceWebSocket$new()}}
\item \href{#method-add_depth_snapshot}{\code{binanceWebSocket$add_depth_snapshot()}}
\item \href{#method-add_trades_snapshot}{\code{binanceWebSocket$add_trades_snapshot()}}
\item \href{#method-add_klines_snapshot}{\code{binanceWebSocket$add_klines_snapshot()}}
\item \href{#method-subscribe}{\code{binanceWebSocket$subscribe()}}
\item \href{#method-unsubscribe}{\code{binanceWebSocket$unsubscribe()}}
\item \href{#method-connect}{\code{binanceWebSocket$connect()}}
\item \href{#method-close}{\code{binanceWebSocket$close()}}
\item \href{#method-get_stream}{\code{binanceWebSocket$get_stream()}}
\item \href{#method-get_info}{\code{binanceWebSocket$get_info()}}
\item \href{#method-clone}{\code{binanceWebSocket$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{binanceWebSocket} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$new(pair = "BTCUSDT", subscription, interval, update_speed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{pair}

\item{\code{subscription}}{subscription}

\item{\code{interval}}{1m}

\item{\code{update_speed}}{1000}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{binanceWebSocket} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_depth_snapshot"></a>}}
\if{latex}{\out{\hypertarget{method-add_depth_snapshot}{}}}
\subsection{Method \code{add_depth_snapshot()}}{
Add Depth
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$add_depth_snapshot(pair, subscription, update_speed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{pair}

\item{\code{subscription}}{subscription}

\item{\code{update_speed}}{1m}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_trades_snapshot"></a>}}
\if{latex}{\out{\hypertarget{method-add_trades_snapshot}{}}}
\subsection{Method \code{add_trades_snapshot()}}{
Add trades
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$add_trades_snapshot(pair, subscription, from, to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{pair}

\item{\code{subscription}}{subscription}

\item{\code{from}}{1000}

\item{\code{to}}{to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_klines_snapshot"></a>}}
\if{latex}{\out{\hypertarget{method-add_klines_snapshot}{}}}
\subsection{Method \code{add_klines_snapshot()}}{
Add Klines
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$add_klines_snapshot(pair, subscription, interval, from, to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{pair}

\item{\code{subscription}}{subscription}

\item{\code{interval}}{1m}

\item{\code{from}}{1000}

\item{\code{to}}{to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subscribe"></a>}}
\if{latex}{\out{\hypertarget{method-subscribe}{}}}
\subsection{Method \code{subscribe()}}{
Subscribe to a stream
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$subscribe(pair, subscription, interval, update_speed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{pair}

\item{\code{subscription}}{subscription}

\item{\code{interval}}{1m}

\item{\code{update_speed}}{1m}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unsubscribe"></a>}}
\if{latex}{\out{\hypertarget{method-unsubscribe}{}}}
\subsection{Method \code{unsubscribe()}}{
Unsubscribe from a stream
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$unsubscribe(pair, subscription, interval, update_speed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{pair}

\item{\code{subscription}}{subscription}

\item{\code{interval}}{1m}

\item{\code{update_speed}}{1m}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect"></a>}}
\if{latex}{\out{\hypertarget{method-connect}{}}}
\subsection{Method \code{connect()}}{
Start the connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$connect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
Close all connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_stream"></a>}}
\if{latex}{\out{\hypertarget{method-get_stream}{}}}
\subsection{Method \code{get_stream()}}{
Unsubscribe from a stream
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$get_stream(pair, subscription, interval, update_speed, id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{pair}

\item{\code{subscription}}{subscription}

\item{\code{interval}}{1m}

\item{\code{update_speed}}{1m}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_info"></a>}}
\if{latex}{\out{\hypertarget{method-get_info}{}}}
\subsection{Method \code{get_info()}}{
Unsubscribe from a stream
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$get_info(pair, subscription, interval, update_speed, id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{pair}

\item{\code{subscription}}{subscription}

\item{\code{interval}}{1m}

\item{\code{update_speed}}{1m}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
